+++ Changelog.rst
@@ -11,6 +11,9 @@ New features
 * For CBC/CFB/OFB, the cipher object has both `IV` and `iv` attributes.
   `new` accepts `IV` as well as `iv` as parameter.
 * Dedicated tests for CBC/CFB/OFB/CTR, including NIST test vectors
+* CCM cipher object has attribute `nonce`.
+* CCM cipher checks if the declared lengths of the associated data and of the
+  message match the actual ones.
 
 Resolved issues
 ---------------